sPlayerId => uiSession
git-svn-id: https://mumble.svn.sourceforge.net/svnroot/mumble/trunk@637 05730e5d-ab1b-0410-a4ac-84af385074fa

+++ src/mumble/Overlay.h
@@ -104,7 +104,7 @@ class Overlay : public QObject {
 			TextLine(const QString &t, quint32 c, int p = -1, Decoration d = None) : qsText(t), iPlayer(p), uiColor(c), dDecor(d) { };
 		};
 
-		typedef QPair<short, QByteArray> UserTexture;
+		typedef QPair<unsigned int, QByteArray> UserTexture;
 		QByteArray qbaMuted, qbaDeafened;
 		QList<TextLine> qlCurrentTexts;
 		QHash<QString, unsigned char *> qhTextures;
